feat: add location data at event

+++ app/src/main/java/ru/practicum/explorewithme/event/dto/EventCreateDto.java
@@ -47,9 +47,9 @@ public class EventCreateDto {
     @AllArgsConstructor
     public static class Location {
 
-        private int lat;
+        private Double lat;
 
-        private int lon;
+        private Double lon;
 
     }
 }
